code structure: add auto recognition for rules
- Add preprocessor `hadolint-discover` which discovers rules DLxxxx
  automatically
- Make all rules take a configuration as argument. A rule can always
  choose to ignore the configuration, or it can adapt given the
  configuration parameters

The preprocessor `hadolint-discover` searches for modules named `DL*.hs`
in subdirectories from where it was called during the preprocessing
step of the build. Then it generates a module which exports an
`autorule` function. This is a fold of all discovered rules (modules
named `DL*.hs`. Given the rule `Hadolint/Rule/Auto.hs`, which triggers
this preprocessor for the `Hadolint/Rule` subdirectory, this results in
automatic detection and inclusion of all rules in that directory named
`DL*.hs`. As a result extending Hadolint with new rules is just as easy
as dropping a new rule module in the `src/Hadolint/Rule` directory and
the corresponding spec file for its tests in the `test/Hadolint/Rule/`
directory.